+++ ddtrace/contrib/internal/fastapi/patch.py
@@ -8,7 +8,7 @@ from wrapt import wrap_function_wrapper as _w
 from ddtrace import Pin
 from ddtrace import config
 from ddtrace.contrib.asgi.middleware import TraceMiddleware
-from ddtrace.contrib.starlette.patch import _trace_background_tasks
+from ddtrace.contrib.internal.starlette.patch import _trace_background_tasks
 from ddtrace.contrib.starlette.patch import traced_handler
 from ddtrace.internal.logger import get_logger
 from ddtrace.internal.schema import schematize_service_name
