+++ ddtrace/contrib/sanic/__init__.py
@@ -62,8 +62,12 @@ required_modules = ["sanic"]
 
 with require_modules(required_modules) as missing_modules:
     if not missing_modules:
-        from .patch import get_version
-        from .patch import patch
-        from .patch import unpatch
+        # Required to allow users to import from `ddtrace.contrib.sanic.patch` directly
+        from . import patch as _  # noqa: F401, I001
+
+        # Expose public methods
+        from ..internal.sanic.patch import get_version
+        from ..internal.sanic.patch import patch
+        from ..internal.sanic.patch import unpatch
 
         __all__ = ["patch", "unpatch", "get_version"]
