+++ ddtrace/contrib/snowflake/__init__.py
@@ -65,8 +65,12 @@ required_modules = ["snowflake.connector"]
 
 with require_modules(required_modules) as missing_modules:
     if not missing_modules:
-        from .patch import get_version
-        from .patch import patch
-        from .patch import unpatch
+        # Required to allow users to import from `ddtrace.contrib.snowflake.patch` directly
+        from . import patch as _  # noqa: F401, I001
+
+        # Expose public methods
+        from ..internal.snowflake.patch import get_version
+        from ..internal.snowflake.patch import patch
+        from ..internal.snowflake.patch import unpatch
 
         __all__ = ["patch", "unpatch", "get_version"]
