+++ ddtrace/contrib/sqlalchemy/__init__.py
@@ -26,9 +26,13 @@ required_modules = ["sqlalchemy", "sqlalchemy.event"]
 
 with require_modules(required_modules) as missing_modules:
     if not missing_modules:
-        from .engine import trace_engine
-        from .patch import get_version
-        from .patch import patch
-        from .patch import unpatch
+        # Required to allow users to import from `ddtrace.contrib.sqlalchemy.patch` directly
+        from . import patch as _  # noqa: F401, I001
+
+        # Expose public methods
+        from ..internal.sqlalchemy.engine import trace_engine
+        from ..internal.sqlalchemy.patch import get_version
+        from ..internal.sqlalchemy.patch import patch
+        from ..internal.sqlalchemy.patch import unpatch
 
         __all__ = ["trace_engine", "patch", "unpatch", "get_version"]
