+++ ddtrace/contrib/starlette/__init__.py
@@ -72,8 +72,12 @@ required_modules = ["starlette"]
 
 with require_modules(required_modules) as missing_modules:
     if not missing_modules:
-        from .patch import get_version
-        from .patch import patch
-        from .patch import unpatch
+        # Required to allow users to import from `ddtrace.contrib.starlette.patch` directly
+        from . import patch as _  # noqa: F401, I001
+
+        # Expose public methods
+        from ..internal.starlette.patch import get_version
+        from ..internal.starlette.patch import patch
+        from ..internal.starlette.patch import unpatch
 
         __all__ = ["patch", "unpatch", "get_version"]
