+++ ddtrace/contrib/structlog/__init__.py
@@ -45,8 +45,12 @@ required_modules = ["structlog"]
 
 with require_modules(required_modules) as missing_modules:
     if not missing_modules:
-        from .patch import get_version
-        from .patch import patch
-        from .patch import unpatch
+        # Required to allow users to import from `ddtrace.contrib.structlog.patch` directly
+        from . import patch as _  # noqa: F401, I001
+
+        # Expose public methods
+        from ..internal.structlog.patch import get_version
+        from ..internal.structlog.patch import patch
+        from ..internal.structlog.patch import unpatch
 
         __all__ = ["patch", "unpatch", "get_version"]
