+++ ddtrace/contrib/subprocess/__init__.py
@@ -26,8 +26,12 @@ required_modules = ["os", "subprocess"]
 
 with require_modules(required_modules) as missing_modules:
     if not missing_modules:
-        from .patch import get_version
-        from .patch import patch
-        from .patch import unpatch
+        # Required to allow users to import from `ddtrace.contrib.subprocess.patch` directly
+        from . import patch as _  # noqa: F401, I001
+
+        # Expose public methods
+        from ..internal.subprocess.patch import get_version
+        from ..internal.subprocess.patch import patch
+        from ..internal.subprocess.patch import unpatch
 
         __all__ = ["patch", "unpatch", "get_version"]
